# swagger/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "YourCoolingPartner API",
    "description": "REST API for the YourCoolingPartner agentic system. Connects Android users with local AC & plumbing technicians using AI-powered search, ranking, and conversational booking — all through a chat interface.",
    "version": "2.0.0",
    "contact": {
      "name": "YourCoolingPartner Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Development Server"
    }
  ],
  "paths": {
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login",
        "description": "Login with mobile number and password. Users must be pre-registered via seeds. No signup endpoint exists.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" },
              "examples": {
                "user_login": {
                  "summary": "Login as a user",
                  "value": {
                    "mobile_number": "03001234567",
                    "password": "password123"
                  }
                },
                "technician_login": {
                  "summary": "Login as a technician",
                  "value": {
                    "mobile_number": "03031234567",
                    "password": "tech123"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Returns JWT token.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/LoginResponse" }
              }
            }
          },
          "401": {
            "description": "Mobile number not registered or incorrect password."
          }
        }
      }
    },
    "/api/chat": {
      "post": {
        "tags": ["AI Agent"],
        "summary": "Chat with AI Agent",
        "description": "Send a message to the Brain Agent. The agent understands Roman Urdu/English and decides what action to take: search technicians, create a job, show bids, create a booking, or ask for clarification. All messages are persisted in the database.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatRequest" },
              "examples": {
                "search": {
                  "summary": "Search for technicians",
                  "value": { "message": "Mujhy Johar Town Lahore ma AC repair chaiye" }
                },
                "partial": {
                  "summary": "Partial info (agent will clarify)",
                  "value": { "message": "Mujhy AC repair chaiye" }
                },
                "schedule": {
                  "summary": "Schedule a job (after searching)",
                  "value": { "message": "Kal subah 10 baje" }
                },
                "check_bids": {
                  "summary": "Check bids on posted job",
                  "value": { "message": "Bids dikhao" }
                },
                "book": {
                  "summary": "Book a technician",
                  "value": { "message": "Bilal ko book karo" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Agent response with action taken.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChatResponse" },
                "examples": {
                  "clarify": {
                    "summary": "Agent asks for more info",
                    "value": {
                      "response": "Aap kis shehar aur ilaqay mein service chahte hain?",
                      "action": "clarify"
                    }
                  },
                  "search_results": {
                    "summary": "Agent returns ranked technicians",
                    "value": {
                      "response": "Yeh lijiye aapke top technicians:\n1. Cool Master AC — ⭐ 4.8 — 03001234567\n2. Bilal AC Wala — ⭐ 4.5 — 03031234567\nKab schedule karna chahte hain?",
                      "action": "search_technicians"
                    }
                  },
                  "job_created": {
                    "summary": "Job posted successfully",
                    "value": {
                      "response": "✅ Aapka job post ho gaya hai! (Job #5)\nTechnicians se bids ka intezaar karein.",
                      "action": "create_job"
                    }
                  },
                  "bids_shown": {
                    "summary": "Bids listed",
                    "value": {
                      "response": "🔔 Job #5 par 2 bids hain:\n1. Bilal AC Wala — Labour: Rs.1200\n2. Kamran Cooling — Labour: Rs.1800\nKisko book karna chahte hain?",
                      "action": "show_bids"
                    }
                  },
                  "booking_done": {
                    "summary": "Booking confirmed",
                    "value": {
                      "response": "✅ Booking confirm ho gayi hai!\n👤 Technician: Bilal AC Wala\n💰 Labour: Rs.1200\nParts replacement charges alag se lagein ge.",
                      "action": "create_booking"
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Missing or invalid JWT token." }
        }
      }
    },
    "/api/chat/history": {
      "get": {
        "tags": ["AI Agent"],
        "summary": "Get Chat History",
        "description": "Returns the full conversation history for the logged-in user. Call this when the app opens to restore the chat view.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Conversation history with all messages.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ChatHistoryResponse" },
                "example": {
                  "conversation_id": 1,
                  "messages": [
                    { "id": 1, "sender": "user", "content": "Mujhy AC repair chaiye Johar Town Lahore", "timestamp": "2026-05-19T10:00:01" },
                    { "id": 2, "sender": "agent", "content": "Yeh lijiye aapke top technicians...", "timestamp": "2026-05-19T10:00:05" },
                    { "id": 3, "sender": "user", "content": "Kal 10 baje", "timestamp": "2026-05-19T10:01:00" },
                    { "id": 4, "sender": "agent", "content": "✅ Job post ho gaya hai!", "timestamp": "2026-05-19T10:01:03" }
                  ]
                }
              }
            }
          },
          "401": { "description": "Missing or invalid JWT token." }
        }
      }
    },
    "/api/jobs": {
      "post": {
        "tags": ["Jobs"],
        "summary": "Create a Job (Manual)",
        "description": "Manually create a job ticket. Jobs are also auto-created by the AI agent when the user provides scheduling info in chat.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/JobCreate" },
              "example": { "city": "Lahore", "town": "Johar Town", "date": "2026-05-20", "time": "10:00 AM" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job created.",
            "content": {
              "application/json": {
                "example": { "message": "Job created", "job_id": 1 }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Jobs"],
        "summary": "Get Jobs",
        "description": "Users see their own jobs. Technicians see all pending jobs available for bidding.",
        "security": [{ "BearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of jobs.",
            "content": {
              "application/json": {
                "example": [
                  { "id": 1, "user_id": 1, "city": "Lahore", "town": "Johar Town", "status": "pending", "date": "2026-05-20", "time": "10:00 AM" }
                ]
              }
            }
          }
        }
      }
    },
    "/api/bids": {
      "post": {
        "tags": ["Bids"],
        "summary": "Submit a Bid",
        "description": "Technicians submit a labour charge bid on an open job. Only users with role 'technician' can submit bids.",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BidSubmit" },
              "example": { "job_id": 1, "amount": 1500.0 }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bid submitted.",
            "content": {
              "application/json": {
                "example": { "message": "Bid submitted", "bid_id": 1 }
              }
            }
          },
          "403": { "description": "Only technicians can submit bids." },
          "404": { "description": "Job not found." }
        }
      }
    },
    "/api/jobs/{job_id}/bids": {
      "get": {
        "tags": ["Bids"],
        "summary": "View Bids on a Job",
        "description": "See all bids placed by technicians on a specific job.",
        "security": [{ "BearerAuth": [] }],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of bids.",
            "content": {
              "application/json": {
                "example": [
                  { "id": 1, "technician_id": 3, "amount": 1500.0 },
                  { "id": 2, "technician_id": 4, "amount": 1200.0 }
                ]
              }
            }
          },
          "404": { "description": "Job not found." }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Get your token from POST /auth/login"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["mobile_number", "password"],
        "properties": {
          "mobile_number": { "type": "string", "example": "03001234567", "description": "11-digit Pakistani mobile number starting with 03" },
          "password": { "type": "string", "example": "password123" }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "access_token": { "type": "string" },
          "token_type": { "type": "string", "example": "bearer" },
          "user": {
            "type": "object",
            "properties": {
              "id": { "type": "integer" },
              "name": { "type": "string" },
              "mobile_number": { "type": "string" },
              "role": { "type": "string" }
            }
          }
        }
      },
      "ChatRequest": {
        "type": "object",
        "required": ["message"],
        "properties": {
          "message": { "type": "string", "example": "Mujhy Johar Town Lahore ma AC repair chaiye" }
        }
      },
      "ChatResponse": {
        "type": "object",
        "properties": {
          "response": { "type": "string", "description": "Agent's reply text" },
          "action": { "type": "string", "enum": ["clarify", "search_technicians", "create_job", "show_bids", "create_booking"], "description": "What action the agent took" }
        }
      },
      "ChatHistoryResponse": {
        "type": "object",
        "properties": {
          "conversation_id": { "type": "integer", "nullable": true },
          "messages": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": { "type": "integer" },
                "sender": { "type": "string", "enum": ["user", "agent"] },
                "content": { "type": "string" },
                "timestamp": { "type": "string" }
              }
            }
          }
        }
      },
      "JobCreate": {
        "type": "object",
        "required": ["city", "town", "date", "time"],
        "properties": {
          "city": { "type": "string", "example": "Lahore" },
          "town": { "type": "string", "example": "Johar Town" },
          "date": { "type": "string", "example": "2026-05-20" },
          "time": { "type": "string", "example": "10:00 AM" }
        }
      },
      "BidSubmit": {
        "type": "object",
        "required": ["job_id", "amount"],
        "properties": {
          "job_id": { "type": "integer", "example": 1 },
          "amount": { "type": "number", "example": 1500.0, "description": "Labour charge in PKR" }
        }
      }
    }
  }
}